Move register_editable_textbox_handlers to a separate module
