+++ traitsui/testing/tester/wx/common_ui_targets.py
@@ -17,8 +17,7 @@ target_class of choice to one of these as the locator_class. For an example,
 see the implementation of range_editor.
 """
 
-from traitsui.testing.tester import command, query
-from traitsui.testing.tester.wx import helpers
+from traitsui.testing.tester.wx import registry_helper
 
 
 class LocatedTextbox:
@@ -44,42 +43,8 @@ class LocatedTextbox:
         registry : TargetRegistry
             The registry being registered to.
         """
-        register_editable_textbox_handlers(
+        registry_helper.register_editable_textbox_handlers(
             registry=registry,
             target_class=cls,
             widget_getter=lambda wrapper: wrapper.target.textbox,
-        )
-
-
-def register_editable_textbox_handlers(registry, target_class, widget_getter):
-    """ Register common interactions for an editable textbox (in Wx)
-
-    Parameters
-    ----------
-    registry : TargetRegistry
-        The registry being registered to.
-    target_class : subclass of type
-        The type of target being wrapped in a UIWrapper on which the
-        interaction will be performed.
-    widget_getter : callable(wrapper: UIWrapper) -> wx.TextCtrl
-        A callable to return a wx.TextCtrl
-    """
-    handlers = [
-        (command.KeySequence,
-            (lambda wrapper, interaction: helpers.key_sequence_text_ctrl(
-                widget_getter(wrapper), interaction, wrapper.delay))),
-        (command.KeyClick,
-            (lambda wrapper, interaction: helpers.key_click_text_ctrl(
-                widget_getter(wrapper), interaction, wrapper.delay))),
-        (command.MouseClick,
-            (lambda wrapper, _: helpers.mouse_click_object(
-                widget_getter(wrapper), wrapper.delay))),
-        (query.DisplayedText,
-            lambda wrapper, _: widget_getter(wrapper).GetValue()),
-    ]
-    for interaction_class, handler in handlers:
-        registry.register_handler(
-            target_class=target_class,
-            interaction_class=interaction_class,
-            handler=handler,
         )
